refactor: move create pat into use case
