+++ src/features/personal_access_token/personal_access_token_feature.py
@@ -6,8 +6,10 @@ from fastapi.responses import JSONResponse, PlainTextResponse
 from authentication.access_control import AccessLevel
 from authentication.authentication import auth_with_jwt
 from authentication.models import PATData, User
-from authentication.personal_access_token import create_personal_access_token
 from common.responses import create_response, responses
+from features.personal_access_token.use_cases.create_personal_access_token_use_case import (
+    create_personal_access_token_use_case,
+)
 from storage.internal.personal_access_tokens import delete_pat, get_users_pats
 
 router = APIRouter(tags=["default", "personal_access_token"], prefix="/token")
@@ -31,7 +33,7 @@ async def new_personal_access_token(
     Returns:
     - str: The generated PAT token
     """
-    return create_personal_access_token(user, scope, time_to_live)
+    return create_personal_access_token_use_case(user, scope, time_to_live)
 
 
 @router.delete("/{token_id}", operation_id="token_delete", response_model=str, responses=responses)
